refactor to class based tiles, and added methods
This commit is necessary in order to create a system that will auto
create each board, in a semi-random system.  Using Classes will allow
this step to be taken.
Classes that were created:
	- Tile
		This is the base class for all future tile classes. This
		creates a
	- Water
	- Island_tile
		class will provide a display, a bool of passable, and a
		bool of has treasure.

classes Water, Island_tile, Whirlpool, and Waterfall.  Refactored the
	- Water
	- Whirlpool
	- Waterfall
		These are various tile that were created for the game.

	- @ship
		This is a hash table that will probably be turned into
		a class in the future with inherited classes that
		players can upgrade and purchase new types of ships

The movement system was refactored. Currently has a bug, that is being
put aside for the moment until the program can fully create boards with
random tiles.