Refactor way that the first level is created
This commit was necessary because previously a global variable '$grid'
was being created in a way that didn't make much sense. It is also a
step in the direction of being able to create levels automatically.
Using small encapsulated methods to create everything in a semi-random
fashion.

The ways that this commit addressed that was by getting a rid of the
global '$grid' in favor of an instance variable '@grid' which lives
inside the Board class.  These variables may be changed around where
that class is 'Level' and the '@level_one' variable is '@board'. This
commit refactors the placement of the individual tiles into a method
called 'create_level_one' which places the tiles that are not water onto
the board.