Add create method on Island class
This commit is necessary to continue down the path of automation of
levels in a semi-random way. Better encapsulation is necessary to
complete this process

This commit creates blocks of island, that will build to form a shape
with a waterfall in the middle.  This will go onto be one of several
shapes that are available to place in each level.  This is the first
step in that process.  The creation of blocks takes in the current grid,
the number of rows, the number of columns, the starting x point, and
the starting y point. It then uses a double for loop to fill in these
columns.