Create method that makes a 'u-shape' island
This commit is necessary to further the process of level automation.  It
is also necessary to provide more encapsulation.  The creation of the
'u-shape' will allow several to be plotted easily. This commit is also
necessary to align the first level with the first levels design, and is
the penultimate step to creating the first level in a semi-random way.

This commit creates a new method that creates a 'u-shaped' island.  It
also places the waterfall in the correct place within that island.  In
order to align the map with the design, the program places three of
these u-shaped islands throughout the map.  The 'u-shape' method takes
in the current grid, a starting x point and a starting y point.

It then adds several blocks together and places the waterfall in the
correct place within the grid, and returns the grid.