Make level creation maps semi-random
This commit was necessary to finish the automatic creation(as opposed to
hard-coded) of a semi-random first level.

This commit addresses that by incorporating a random number generator
in place of the start x and start y inputs for the 'u-shape' island
creator. This commit addresses something out of scope which is that it
creates another file to keep the creation of tiles, and all classes that
inherit the Tile class.