Fix and refactor movement process, bug
The movement function and process wasn't working, and before when it did
work is wasn't very readable, and was a bunch of nested if statements
that didn't make much sense.  The movement had stopped working because
of a previous attempted refactoring.

The changes in this commit address that by using some encapsulation of
the different parts of how the movement function works.  It now takes in
the input key press, takes that and makes a potential move.  This
potential move is passed to the check move function which checks for out
of bounds, and then checks to see if the path is passable.  This check
move function then returns a bool value. If the bool that is returned is
true, the potential moves are accounted for and the ships coordinates
are changed.