Add scroll & fog of war functionality
This commit probably should have been 2 commits.  These changes are to
add a scrolling screen that shows a limited map, and to add a fog of
war.

= Scrolling screen is done by not dislaying tiles that are not in the
current screen, this is set by the vertical max and min & horizontal max
and min variables and comparing them to the tile.

- Scroll necessity is checked then applied to the maxes and mins

- fog of war was done by adding a class method that will change the
value of revealed class variable.  This then changes a switch in the
display function which will either display the tile with it's color or
display it as white.

- created several new variables for this on @ship, and on the tiles
themselves.